+++ src/admin.rs
@@ -367,11 +367,12 @@ impl AdminState {
                 // Create rows for each alias
                 let rows: Vec<Row<'_>> = data
                     .iter()
-                    .map(|(host, connections)| {
+                    .map(|(host, (connections, conns_per_min))| {
                         let len = connections.len() as u16;
                         let (peers, users): (Vec<_>, Vec<_>) = connections.iter().unzip();
                         Row::new(vec![
                             host.clone(),
+                            conns_per_min.to_string(),
                             users.iter().join("\n"),
                             peers.iter().map(to_socket_addr_string).join("\n"),
                         ])
@@ -380,11 +381,12 @@ impl AdminState {
                     .collect();
                 let constraints = [
                     Constraint::Min(25),
+                    Constraint::Length(7),
                     Constraint::Length(50),
                     Constraint::Length(47),
                 ];
-                let header =
-                    Row::new(["Alias", "User(s)", "Peer(s)"]).add_modifier(Modifier::UNDERLINED);
+                let header = Row::new(["Alias", "Con/min", "User(s)", "Peer(s)"])
+                    .add_modifier(Modifier::UNDERLINED);
                 let title =
                     Block::new().title(Line::from("SSH services".fg(color).bold()).centered());
                 Table::new(rows, constraints)
@@ -400,11 +402,12 @@ impl AdminState {
                 // Create rows for each socket or alias
                 let rows: Vec<Row<'_>> = data
                     .iter()
-                    .map(|(port, connections)| {
+                    .map(|(port, (connections, conns_per_min))| {
                         let len = connections.len() as u16;
                         let (peers, users): (Vec<_>, Vec<_>) = connections.iter().unzip();
                         Row::new(vec![
                             port.to_string(),
+                            conns_per_min.to_string(),
                             users.iter().join("\n"),
                             peers.iter().map(to_socket_addr_string).join("\n"),
                         ])
@@ -413,11 +416,12 @@ impl AdminState {
                     .collect();
                 let constraints = [
                     Constraint::Length(5),
+                    Constraint::Length(7),
                     Constraint::Length(50),
                     Constraint::Length(47),
                 ];
-                let header =
-                    Row::new(["Port", "User(s)", "Peer(s)"]).add_modifier(Modifier::UNDERLINED);
+                let header = Row::new(["Port", "Con/min", "User(s)", "Peer(s)"])
+                    .add_modifier(Modifier::UNDERLINED);
                 let title =
                     Block::new().title(Line::from("TCP services".fg(color).bold()).centered());
                 Table::new(rows, constraints)
@@ -433,11 +437,12 @@ impl AdminState {
                 // Create rows for each socket or alias
                 let rows: Vec<Row<'_>> = data
                     .iter()
-                    .map(|(TcpAlias(alias, port), connections)| {
+                    .map(|(TcpAlias(alias, port), (connections, conns_per_min))| {
                         let len = connections.len() as u16;
                         let (peers, users): (Vec<_>, Vec<_>) = connections.iter().unzip();
                         Row::new(vec![
                             format!("{}:{}", alias, port),
+                            conns_per_min.to_string(),
                             users.iter().join("\n"),
                             peers.iter().map(to_socket_addr_string).join("\n"),
                         ])
@@ -446,11 +451,12 @@ impl AdminState {
                     .collect();
                 let constraints = [
                     Constraint::Min(25),
+                    Constraint::Length(7),
                     Constraint::Length(50),
                     Constraint::Length(47),
                 ];
-                let header =
-                    Row::new(["Alias", "User(s)", "Peer(s)"]).add_modifier(Modifier::UNDERLINED);
+                let header = Row::new(["Alias", "Con/min", "User(s)", "Peer(s)"])
+                    .add_modifier(Modifier::UNDERLINED);
                 let title =
                     Block::new().title(Line::from("Alias services".fg(color).bold()).centered());
                 Table::new(rows, constraints)
@@ -814,7 +820,7 @@ impl AdminInterface {
                                 .unwrap()
                                 .values()
                                 .nth(row)
-                                .map(|value| value.values().cloned().collect()),
+                                .map(|value| value.0.values().cloned().collect()),
                             Tab::Tcp => interface
                                 .state
                                 .server
@@ -823,7 +829,7 @@ impl AdminInterface {
                                 .unwrap()
                                 .values()
                                 .nth(row)
-                                .map(|value| value.values().cloned().collect()),
+                                .map(|value| value.0.values().cloned().collect()),
                             Tab::Alias => interface
                                 .state
                                 .server
@@ -832,7 +838,7 @@ impl AdminInterface {
                                 .unwrap()
                                 .values()
                                 .nth(row)
-                                .map(|value| value.values().cloned().collect()),
+                                .map(|value| value.0.values().cloned().collect()),
                         };
                         match users {
                             None => {